test(edit patient): fix tests related to implementing Edit Patient
