+++ src/__tests__/patients/patient-slice.test.ts
@@ -1,9 +1,16 @@
+import '../../__mocks__/matchMediaMock'
 import { AnyAction } from 'redux'
 import { mocked } from 'ts-jest/utils'
+import { createMemoryHistory } from 'history'
+import * as components from '@hospitalrun/components'
+
 import patient, {
   getPatientStart,
   getPatientSuccess,
   fetchPatient,
+  createPatientStart,
+  createPatientSuccess,
+  createPatient,
   updatePatientStart,
   updatePatientSuccess,
   updatePatient,
@@ -16,7 +23,7 @@ describe('patients slice', () => {
     jest.resetAllMocks()
   })
 
-  describe('patients reducer', () => {
+  describe('patient reducer', () => {
     it('should create the proper initial state with empty patients array', () => {
       const patientStore = patient(undefined, {} as AnyAction)
       expect(patientStore.isLoading).toBeFalsy()
@@ -51,6 +58,22 @@ describe('patients slice', () => {
       expect(patientStore.patient).toEqual(expectedPatient)
     })
 
+    it('should handle the CREATE_PATIENT_START action', () => {
+      const patientsStore = patient(undefined, {
+        type: createPatientStart.type,
+      })
+
+      expect(patientsStore.isLoading).toBeTruthy()
+    })
+
+    it('should handle the CREATE_PATIENT_SUCCESS actions', () => {
+      const patientsStore = patient(undefined, {
+        type: createPatientSuccess.type,
+      })
+
+      expect(patientsStore.isLoading).toBeFalsy()
+    })
+
     it('should handle the UPDATE_PATIENT_START action', () => {
       const patientStore = patient(undefined, {
         type: updatePatientStart.type,
@@ -77,6 +100,90 @@ describe('patients slice', () => {
 
       expect(patientStore.isLoading).toBeFalsy()
       expect(patientStore.patient).toEqual(expectedPatient)
+    })
+  })
+
+  describe('createPatient()', () => {
+    it('should dispatch the CREATE_PATIENT_START action', async () => {
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+      const expectedPatient = {
+        id: 'id',
+      } as Patient
+
+      await createPatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
+
+      expect(dispatch).toHaveBeenCalledWith({ type: createPatientStart.type })
+    })
+
+    it('should call the PatientRepository save method with the correct patient', async () => {
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+      jest.spyOn(PatientRepository, 'save')
+      const expectedPatient = {
+        id: 'id',
+      } as Patient
+
+      await createPatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
+
+      expect(PatientRepository.save).toHaveBeenCalledWith(expectedPatient)
+    })
+
+    it('should dispatch the CREATE_PATIENT_SUCCESS action', async () => {
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+      const mockedPatientRepository = mocked(PatientRepository, true)
+      mockedPatientRepository.save.mockResolvedValue({ id: '12345' } as Patient)
+      const expectedPatient = {
+        id: 'id',
+      } as Patient
+
+      await createPatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
+
+      expect(dispatch).toHaveBeenCalledWith({ type: createPatientSuccess.type })
+    })
+
+    it('should navigate to the /patients/:id where id is the new patient id', async () => {
+      const expectedPatientId = '12345'
+      const mockedPatientRepository = mocked(PatientRepository, true)
+      mockedPatientRepository.save.mockResolvedValue({ id: expectedPatientId } as Patient)
+      const history = createMemoryHistory()
+
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+      const expectedPatient = {} as Patient
+
+      await createPatient(expectedPatient, history)(dispatch, getState, null)
+
+      expect(history.entries[1].pathname).toEqual(`/patients/${expectedPatientId}`)
+    })
+
+    it('should call the Toaster function with the correct data', async () => {
+      jest.spyOn(components, 'Toast')
+      const expectedPatientId = '12345'
+      const expectedGivenName = 'given'
+      const expectedFamilyName = 'family'
+      const expectedSuffix = 'suffix'
+      const expectedPatient = {
+        id: expectedPatientId,
+        givenName: expectedGivenName,
+        familyName: expectedFamilyName,
+        suffix: expectedSuffix,
+      } as Patient
+      const mockedPatientRepository = mocked(PatientRepository, true)
+      mockedPatientRepository.save.mockResolvedValue(expectedPatient)
+      const mockedComponents = mocked(components, true)
+      const history = createMemoryHistory()
+      const dispatch = jest.fn()
+      const getState = jest.fn()
+
+      await createPatient(expectedPatient, history)(dispatch, getState, null)
+
+      expect(mockedComponents.Toast).toHaveBeenCalledWith(
+        'success',
+        'Success!',
+        `Successfully created patient ${expectedGivenName} ${expectedFamilyName} ${expectedSuffix}`,
+      )
     })
   })
 
@@ -130,6 +237,7 @@ describe('patients slice', () => {
     })
   })
 
+  // should check for the Toast
   describe('update patient', () => {
     it('should dispatch the UPDATE_PATIENT_START action', async () => {
       const dispatch = jest.fn()
@@ -140,7 +248,7 @@ describe('patients slice', () => {
       const mockedPatientRepository = mocked(PatientRepository, true)
       mockedPatientRepository.saveOrUpdate.mockResolvedValue(expectedPatient)
 
-      await updatePatient(expectedPatient)(dispatch, getState, null)
+      await updatePatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
 
       expect(dispatch).toHaveBeenCalledWith({ type: updatePatientStart.type })
     })
@@ -154,7 +262,7 @@ describe('patients slice', () => {
       const mockedPatientRepository = mocked(PatientRepository, true)
       mockedPatientRepository.saveOrUpdate.mockResolvedValue(expectedPatient)
 
-      await updatePatient(expectedPatient)(dispatch, getState, null)
+      await updatePatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
 
       expect(PatientRepository.saveOrUpdate).toHaveBeenCalledWith(expectedPatient)
     })
@@ -168,7 +276,7 @@ describe('patients slice', () => {
       const mockedPatientRepository = mocked(PatientRepository, true)
       mockedPatientRepository.saveOrUpdate.mockResolvedValue(expectedPatient)
 
-      await updatePatient(expectedPatient)(dispatch, getState, null)
+      await updatePatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
 
       expect(dispatch).toHaveBeenCalledWith({
         type: updatePatientSuccess.type,
