+++ src/__tests__/patients/patients-slice.test.ts
@@ -1,14 +1,9 @@
 import '../../__mocks__/matchMediaMock'
 import { AnyAction } from 'redux'
-import { createMemoryHistory } from 'history'
 import { mocked } from 'ts-jest/utils'
-import * as components from '@hospitalrun/components'
 import patients, {
   getPatientsStart,
   getAllPatientsSuccess,
-  createPatientStart,
-  createPatientSuccess,
-  createPatient,
   searchPatients,
 } from '../../patients/patients-slice'
 import Patient from '../../model/Patient'
@@ -26,22 +21,6 @@ describe('patients slice', () => {
       expect(patientsStore.patients).toHaveLength(0)
     })
 
-    it('should handle the CREATE_PATIENT_START action', () => {
-      const patientsStore = patients(undefined, {
-        type: createPatientStart.type,
-      })
-
-      expect(patientsStore.isLoading).toBeTruthy()
-    })
-
-    it('should handle the CREATE_PATIENT_SUCCESS actions', () => {
-      const patientsStore = patients(undefined, {
-        type: createPatientSuccess.type,
-      })
-
-      expect(patientsStore.isLoading).toBeFalsy()
-    })
-
     it('should handle the GET_ALL_PATIENTS_SUCCESS action', () => {
       const expectedPatients = [{ id: '1234' }]
       const patientsStore = patients(undefined, {
@@ -51,90 +30,6 @@ describe('patients slice', () => {
 
       expect(patientsStore.isLoading).toBeFalsy()
       expect(patientsStore.patients).toEqual(expectedPatients)
-    })
-  })
-
-  describe('createPatient()', () => {
-    it('should dispatch the CREATE_PATIENT_START action', async () => {
-      const dispatch = jest.fn()
-      const getState = jest.fn()
-      const expectedPatient = {
-        id: 'id',
-      } as Patient
-
-      await createPatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
-
-      expect(dispatch).toHaveBeenCalledWith({ type: createPatientStart.type })
-    })
-
-    it('should call the PatientRepository save method with the correct patient', async () => {
-      const dispatch = jest.fn()
-      const getState = jest.fn()
-      jest.spyOn(PatientRepository, 'save')
-      const expectedPatient = {
-        id: 'id',
-      } as Patient
-
-      await createPatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
-
-      expect(PatientRepository.save).toHaveBeenCalledWith(expectedPatient)
-    })
-
-    it('should dispatch the CREATE_PATIENT_SUCCESS action', async () => {
-      const dispatch = jest.fn()
-      const getState = jest.fn()
-      const mockedPatientRepository = mocked(PatientRepository, true)
-      mockedPatientRepository.save.mockResolvedValue({ id: '12345' } as Patient)
-      const expectedPatient = {
-        id: 'id',
-      } as Patient
-
-      await createPatient(expectedPatient, createMemoryHistory())(dispatch, getState, null)
-
-      expect(dispatch).toHaveBeenCalledWith({ type: createPatientSuccess.type })
-    })
-
-    it('should navigate to the /patients/:id where id is the new patient id', async () => {
-      const expectedPatientId = '12345'
-      const mockedPatientRepository = mocked(PatientRepository, true)
-      mockedPatientRepository.save.mockResolvedValue({ id: expectedPatientId } as Patient)
-      const history = createMemoryHistory()
-
-      const dispatch = jest.fn()
-      const getState = jest.fn()
-      const expectedPatient = {} as Patient
-
-      await createPatient(expectedPatient, history)(dispatch, getState, null)
-
-      expect(history.entries[1].pathname).toEqual(`/patients/${expectedPatientId}`)
-    })
-
-    it('should call the Toaster function with the correct data', async () => {
-      jest.spyOn(components, 'Toast')
-      const expectedPatientId = '12345'
-      const expectedGivenName = 'given'
-      const expectedFamilyName = 'family'
-      const expectedSuffix = 'suffix'
-      const expectedPatient = {
-        id: expectedPatientId,
-        givenName: expectedGivenName,
-        familyName: expectedFamilyName,
-        suffix: expectedSuffix,
-      } as Patient
-      const mockedPatientRepository = mocked(PatientRepository, true)
-      mockedPatientRepository.save.mockResolvedValue(expectedPatient)
-      const mockedComponents = mocked(components, true)
-      const history = createMemoryHistory()
-      const dispatch = jest.fn()
-      const getState = jest.fn()
-
-      await createPatient(expectedPatient, history)(dispatch, getState, null)
-
-      expect(mockedComponents.Toast).toHaveBeenCalledWith(
-        'success',
-        'Success!',
-        `Successfully created patient ${expectedGivenName} ${expectedFamilyName} ${expectedSuffix}`,
-      )
     })
   })
