+++ src/__tests__/patients/view/ViewPatient.test.tsx
@@ -6,7 +6,7 @@ import { mocked } from 'ts-jest/utils'
 import { act } from 'react-dom/test-utils'
 import { Route, Router } from 'react-router-dom'
 import { TabsHeader, Tab } from '@hospitalrun/components'
-import GeneralInformation from 'patients/view/GeneralInformation'
+import GeneralInformation from 'patients/GeneralInformation'
 import { createMemoryHistory } from 'history'
 import RelatedPersonTab from 'patients/related-persons/RelatedPersonTab'
 import Patient from '../../../model/Patient'
