+++ src/patients/edit/EditPatient.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { useHistory, useParams } from 'react-router'
+import { useHistory, useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { useDispatch, useSelector } from 'react-redux'
 import { Spinner } from '@hospitalrun/components'
@@ -45,7 +45,7 @@ const EditPatient = () => {
   }, [id, dispatch])
 
   const onCancel = () => {
-    history.goBack()
+    history.push(`/patients/${id}`)
   }
 
   const onSave = () => {
